refactor(db): sweep naming over the db models, root files and db-mock
`ContentType.ts` held one export, `ContentTypes`, and no type of that name for
the file to have been named after — unlike the enum-plus-values files it
resembles, where both exist. The file takes its single export's name.

`createMockDb.test.ts` built its freshly-migrated database as `newPGlite`,
where `new*` is this repo's prefix for an update handler's incoming value; it
is the migrated half of the pair the snapshot is compared against.
