+++ packages/db/src/services/azure/search/filtersToClauses.test.ts
@@ -1,4 +1,4 @@
-import { ContentTypes } from "#src/models/ContentType";
+import { ContentTypes } from "#src/models/ContentTypes";
 import { filtersToClauses } from "#src/services/azure/search/filtersToClauses";
 import { BinaryOperator, CompositeKeyPropertyNames, SearchOperator } from "@esposter/azure";
 import {
